Moved Node class to a separate java file and created linkedlist package
Moved Node class outside the LinkedList.java file to a new Node.java
file, and moved both files to a new package called "linkedlist".
