+++ src/exercise/Main.java
@@ -1,5 +1,7 @@
 package exercise;
 
+import exercise.linkedlist.LinkedList;
+
 public class Main {
 
 	public static void main(String[] args) {
