+++ src/exercise/linkedlist/Node.java
@@ -1,0 +1,11 @@
+package exercise.linkedlist;
+
+class Node {
+	int elem;
+	Node next;
+
+	Node(int i, Node n) {
+		elem = i;
+		next = n;
+	}
+}
